backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-dotenv==1.0.1
anthropic==0.40.0
chromadb==0.5.23
sentence-transformers==3.3.1
langgraph==0.2.60
langchain-core==0.3.28
Pillow==11.0.0
httpx==0.28.1
pydantic==2.10.3
